Refactor build script
